Update onSignIn event handler to check site exists
Added logic to onSignIn event handler to check if the logged in user's
id does not match any of the _owner ids for all the sites returned, then
that means the user does not have a site. If there is match, show the
user the dashboard view, else show the user the createSiteview.